backend/requirements.txt
FastAPI==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
openai>=1.0.0
python-dotenv>=1.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.9.0
numpy>=1.26.0
pdf2image>=1.17.0
requests>=2.31.0
